Refactor, add method to get variant

+++ Feature.Library/FlagVariant.cs
@@ -13,6 +13,6 @@
         /// <summary>
         /// The percentage en
         /// </summary>
-        public int Percentage { get; set; }
+        public double Percentage { get; set; }
     }
 }
